# Models/TollRoad.js
const mongoose = require('mongoose');
const Crud = require('./../Classes/Crud');
const schemas = {
  TollRoad: {
    name : {type: String},
    _wayPoints : [{type: mongoose.Schema.Types.ObjectId, ref: 'WayPoint'}],
    createdAt: {type: Date, default: Date.now()},
    updatedAt: {type: Date}
  }
};

class TollRoad extends Crud {

  constructor() {
    super();
    this.TollRoad = mongoose.model('TollRoad', new mongoose.Schema(schemas.TollRoad));
  }

  create(tollRoad) { return super._create(this.TollRoad, tollRoad); }

  read(_id, limit, skip) { return super._read(this.TollRoad, _id, '_wayPoints', limit, skip); }

  update(_id, changes) { return super._update(this.TollRoad, _id, schemas.TollRoad, changes); }

  remove(_id) { return super._remove(this.TollRoad, _id); }

  findOlder(timestamp, Models) {
    return Promise.resolve({msg: 'in developing...'});
  }

  addWayPoint(_id, _wayPoint) {
    return new Promise((resolve, reject) => {
      this.read(_id)
      .then(road => {
        road._wayPoints.push(_wayPoint);
        road.updatedAt = Date.now();
        return road.save();
      })
      .then(resolve)
      .catch(reject);
    })
  }
}

module.exports = TollRoad;